backend/requirements.txt
# FastAPI
fastapi==0.109.0              # Webフレームワーク本体
uvicorn[standard]==0.27.0     # FastAPIを動かすASGIサーバー

# Database
sqlalchemy==2.0.25            # ORM（オブジェクト関係マッピング）- SQL injectionを防ぐ
alembic==1.13.1               # データベースマイグレーションツール
psycopg2-binary==2.9.9        # PostgreSQL接続ドライバ

# Validation
pydantic==2.5.3               # データ検証ライブラリ（FastAPIと統合）
pydantic-settings==2.1.0      # 環境変数を読み込むための拡張
email-validator==2.1.0        # メールアドレス形式の検証

# Authentication
python-jose[cryptography]==3.3.0  # JWT（JSON Web Token）の生成・検証
passlib[bcrypt]==1.7.4            # パスワードハッシュ化（bcryptアルゴリズム）
bcrypt==4.1.3                     # bcrypt 5.xは passlib 1.7.4 と互換性なし
python-multipart==0.0.6           # フォームデータのパース（ファイルアップロード対応）

# Redis (for session/cache)
redis==5.0.1                  # セッション・キャッシュ管理用

# Testing
pytest==7.4.4                 # テストフレームワーク
httpx==0.26.0                 # HTTPクライアント（APIテスト用）
pytest-asyncio==0.23.3        # 非同期テスト対応
